// src/utils/enums.ts
export enum typesChallenge {
  personal = 3,
  command = 2,
  common = 1
}

export enum stageRegistration {
  email = 'email',
  password = 'password',
  phone = 'phone',
  birthday = 'birthday',
  gender = 'gender',
  userName = 'userName',
  surname = 'surname',
  platform = 'platform',
  privatePlatform = 'privatePlatform',
  photo = 'photo',
  tracker = 'tracker'
}
export enum stageCreatingChallenge {
  customers = 'customers',
  platform = 'platform',
  type = 'typeChallenge',
  target = 'target',
  data = 'data',
  title = 'titleChallenge',
  description = 'descriptionChallenge',
  teams = 'teams',
  finally = 'finally',
  lecture = 'lecture'
}
export enum stageAccessRecovery {
  email = 'email',
  password = 'password'
}

export enum typeImage {
  avatars = 'avatars',
  challenges = 'challenges',
  lessons = 'lessons',
  news = 'news'
}

export enum formatConsultation {
  video = 'video',
  audio = 'audio'
}

export enum PEDOMETERS {
  googleFit = 1,
  healthBalance = 2
}